Modif css/ ajout de vérif sécurité formulaires php

+++ admin_index.php
@@ -10,31 +10,26 @@ if  (!Auth::isLogged()){
     header ('Location: ./index.php?action=authentificationFailed');
 }
 
-
+if(isset($_GET['action'])){
  
 				// si titre, contenu de poste et extrait renseignés, alors on crée un nouveau poste et on redirige
+
 	if ($_GET['action']=="new_post" && !empty($_POST['post_title']) && !empty($_POST['create_post']) && !empty ($_POST['create_extract'])){
 			add_post();
 			header ("Location: index.php");
 	}
+
 	/* --------Si action = create, on redirige vers la vue de création--------------*/
 
 	 if ($_GET['action']=="create"){
 	 		require ('view/backend/create_view.php');
 
-
 	 }
 
-	 	/* --------Si action = read, on récup tous les posts et on redirige vers la vue de lecture--------------*/
-
-	if ($_GET['action']=="read"){
-			$posts=get_all_posts();
-			require('view/backend/read_view.php');
-	}
 
 		/* --------Si action = update, on recup le post et on redirige vers la vue de création--------------*/
 
-		if ($_GET['action']=="update"){
+		if ($_GET['action']=="update" && !empty($_GET['id']) && $_GET['id']>0){
 			$post=getPost($_GET['id']);
 			require('view/backend/update_view.php');
 	}
@@ -49,10 +44,10 @@ if  (!Auth::isLogged()){
 
 		/* --------Si action = delete_post, on recup tous le ^post sélectionné de via checkbox on supprime et redirige--------------*/
 
-if ($_GET['action']=="delete_post"){
+if ($_GET['action']=="delete_post" && !empty($_POST['postId']) && $_POST['postId']>0){
 			$posts=get_all_posts();
-		foreach ($_POST['postId'] as $valeur){ // on parcoure le tableau id sélectionné de checkbox
-			 delete($valeur);
+				foreach ($_POST['postId'] as $valeur){ // on parcoure le tableau id sélectionné de checkbox
+			 		delete($valeur);
 			}
 			$posts=get_all_posts();
 
@@ -62,7 +57,7 @@ if ($_GET['action']=="delete_post"){
 
 		/* --------Si action = deleteComm, on recup le comm, le supprime et on redirige vers le post--------------*/
 
-if ($_GET['action']=="deleteComm"){
+if ($_GET['action']=="deleteComm" && !empty($_GET['comm_id']) && $_GET['comm_id']>0 && !empty($_GET['id']) && $_GET['id']>0){
 			
 			$delete=deleteComm($_GET['comm_id']);
 			post($_GET['id']);
@@ -70,7 +65,7 @@ if ($_GET['action']=="deleteComm"){
 	}
 			/* --------Si action = editPost, on édite et on redirige vers la vue du post--------------*/
 
-if ($_GET['action']=="edit_post"){
+if ($_GET['action']=="edit_post" && !empty($_GET['id']) && $_GET['id']>0){
 
 				$edit=edit($_GET['id'],$_POST['title'],$_POST['post_content'],$_POST['extract']);
 				$post=getPost($_GET['id']);
@@ -81,7 +76,7 @@ if ($_GET['action']=="edit_post"){
 
 			/* --------Si action = edti_comment, on recup le comm et on redirige vers la vue du post--------------*/
 
-if ($_GET['action']=="edit_comment"){
+if ($_GET['action']=="edit_comment" && !empty($_GET['comm_id']) && $_GET['comm_id']>0){
 
 				$edit=editComment($_GET['comm_id'],$_POST['author'],$_POST['comment']);
 				$default=defaultAlert($_GET['comm_id']);
@@ -92,7 +87,7 @@ if ($_GET['action']=="edit_comment"){
 	}
 		/* --------Si action = post_view, on récup le post et redirige--------------*/
 
-	if ($_GET['action']=="post_view"){
+	if ($_GET['action']=="post_view" && !empty($_GET['id']) && $_GET['id']>0){
 
 				post($_GET['id']);
 		
@@ -109,3 +104,11 @@ if ($_GET['action']=="edit_comment"){
 
 	 		require ('view/backend/admin_default.php');
 	}
+
+}
+else{
+				$posts=get_all_posts();
+				$comments=get_all_comments();
+		 		require ('view/backend/admin_default.php');
+
+}
